Cart inserts new cart. Working on cart items. Reviews is fixed

+++ src/main/java/com/revature/eMarket/daos/ProductDAO.java
@@ -34,38 +34,6 @@ public class ProductDAO implements CrudDAO<Product>{
         throw new UnsupportedOperationException("Unimplemented method 'findById'");
     }
 
-//    @Override
-//    public Optional<CartItem> findById(String id) {
-//
-//        try(Connection conn = ConnectionFactory.getInstance().getConnection()) {
-//            String sql = "SELECT * FROM products WHERE id = ?";
-//            try(PreparedStatement ps = conn.prepareStatement(sql)) {
-//                ps.setString(1, id);
-//
-//                try (ResultSet rs = ps.executeQuery()) {
-//                    if (rs.next()) {
-//                        Product prod = new Product();
-//                        prod.setId(rs.getString("id"));
-//                        prod.setName(rs.getString("name"));
-//                        prod.setPrice(rs.getInt("price"));
-//                        prod.setStock(rs.getString("stock"));
-//                        prod.setCategory_id(rs.getString("category_id"));
-//                        return prod;
-//                    }
-//                }
-//            }
-//
-//        }catch (SQLException e) {
-//            throw new RuntimeException("Unable to access the database. Debug");
-//        }catch(ClassNotFoundException e) {
-//            throw new RuntimeException("Can't find application. Debug");
-//        }catch (IOException e) {
-//            throw new RuntimeException("Unable to load JDBC. Debug");
-//        }
-//
-//        return null;
-//    }
-
     @Override
     public List<Product> findAll() {
         List<Product> storeList = new ArrayList<>();
@@ -84,7 +52,7 @@ public class ProductDAO implements CrudDAO<Product>{
                         Product prod = new Product();
                         prod.setId(rs.getString("id"));
                         prod.setName(rs.getString("name"));
-                        prod.setPrice(rs.getInt("price"));
+                        prod.setPrice(rs.getFloat("price"));
                         prod.setStock(rs.getString("stock"));
                         prod.setCategory_id(rs.getString("category_id"));
                         storeList.add(prod);
@@ -116,7 +84,7 @@ public class ProductDAO implements CrudDAO<Product>{
                         Product prod = new Product();
                         prod.setId(rs.getString("id"));
                         prod.setName(rs.getString("name"));
-                        prod.setPrice(rs.getInt("price"));
+                        prod.setPrice(rs.getFloat("price"));
                         prod.setStock(rs.getString("stock"));
                         prod.setCategory_id(rs.getString("category_id"));
                         return prod;
